src/main/java/com/unmeshc/security/services/UserServiceImpl.java
package com.unmeshc.security.services;

import com.unmeshc.security.domain.Role;
import com.unmeshc.security.domain.User;
import com.unmeshc.security.repositories.RoleRepository;
import com.unmeshc.security.repositories.UserRepository;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;

/**
 * Created by uc on 10/9/2019
 */
@Service
public class UserServiceImpl implements UserService {

    private final UserRepository userRepository;
    private final RoleRepository roleRepository;
    private final PasswordEncoder passwordEncoder;

    public UserServiceImpl(UserRepository userRepository,
                           RoleRepository roleRepository,
                           PasswordEncoder passwordEncoder) {
        this.userRepository = userRepository;
        this.roleRepository = roleRepository;
        this.passwordEncoder = passwordEncoder;
    }

    @Override
    public User findUserByEmail(String email) {
        return userRepository.findByEmail(email).orElse(null);
    }

    @Override
    public User saveUser(User user) {
        user.setPassword(passwordEncoder.encode(user.getPassword()));
        user.setActive(true);
        Role role = roleRepository.findByName("ADMIN").get();
        user.getRoles().add(role);

        return userRepository.save(user);
    }
}
